--- schemas/index.js
@@ -1,4 +1,0 @@
-const joiSchemas = require('./contactsJoi');
-const mongoSchemas = require('./contactsMongo');
-
-module.exports = { joiSchemas, mongoSchemas };
